Clean Up
I cleaned up some of the css and html in my code, though not much. I
changed the formatting of my settings page hoping there was some way to
make it dynamically show up on the home screen instead of in a new
webpage, other than that, not much else has changed for today.