Updated Settings Pages
Based on critiques from Monday I added stuff we were missing, like an
option to change our password, a button to save user-profile
information, and now the files in the delete menu actually disappear
when clicked.

 I also added some text reminding our  professor that he isn’t
requiring us to have a backend for our app. The other groups were able
to get almost fully functioning products, but that doesn’t mean we
should be required to as well.